Have a basic application.
The React will interact with the backend node server using express where it will process
the data from the csv and txt file to be displayed for the frontend. In order to generate
the data the python script needs to be run which will enter the data parsed using selenium
to the csv file. Possibly will change this such that socket programming is implemented or even just
change the backend to django so I don't need to run the script seperately and can just have a button
that activates the parsing of selenium.